Band Name Generator
The Band Name Generator is a fun and interactive Python script designed to create unique band names based on user input. By simply providing the name of the city you grew up in and your pet's name, the program generates a catchy and personalized band name that you can imagine performing on stage!